Dataset refactoring
Extracted the handler to get table columns
