+++ backend/dataall/aws/handlers/glue.py
@@ -6,7 +6,6 @@ from .service_handlers import Worker
 from .sts import SessionHelper
 from ... import db
 from ...db import models
-from dataall.modules.datasets.services.dataset_table import DatasetTableService
 
 log = logging.getLogger('aws:glue')
 
@@ -505,35 +504,6 @@ class Glue:
                 log.error('Failed to update crawler %s', e)
             else:
                 raise e
-
-
-    @staticmethod
-    @Worker.handler('glue.table.columns')
-    def get_table_columns(engine, task: models.Task):
-        with engine.scoped_session() as session:
-            dataset_table: models.DatasetTable = session.query(models.DatasetTable).get(
-                task.targetUri
-            )
-            aws = SessionHelper.remote_session(dataset_table.AWSAccountId)
-            glue_client = aws.client('glue', region_name=dataset_table.region)
-            glue_table = {}
-            try:
-                glue_table = glue_client.get_table(
-                    CatalogId=dataset_table.AWSAccountId,
-                    DatabaseName=dataset_table.GlueDatabaseName,
-                    Name=dataset_table.name,
-                )
-            except glue_client.exceptions.ClientError as e:
-                log.error(
-                    f'Failed to get table aws://{dataset_table.AWSAccountId}'
-                    f'//{dataset_table.GlueDatabaseName}'
-                    f'//{dataset_table.name} due to: '
-                    f'{e}'
-                )
-            DatasetTableService.sync_table_columns(
-                session, dataset_table, glue_table['Table']
-            )
-        return True
 
     @staticmethod
     @Worker.handler(path='glue.job.runs')
